feat: add & check for role in DELETE

+++ src/routes/product.ts
@@ -76,6 +76,8 @@ productRouter
   .delete(authenticateToken, async (_req, res, next) => {
     const { id } = _req.params;
 
+    if (_req.user?.role !== 'admin') return res.sendStatus(401);
+
     try {
       const result = await collections.products?.deleteOne({ _id: new ObjectId(id) });
